Implement ApiResponseDeferredCallAdapter and add it to the factory
